The composite mutations are dissgregated from simple and factories
